refactor: display costs as currency
Before this commit, the game displayed costs as plain numbers. With this update, a currency system has been implemented to display these values in a more readable format.
The currency system includes abbreviations such as k (thousand), m (million), b (billion), t (trillion), aa, ab, and so on, based on the current value. This makes it easier for the player to understand the magnitude of the numbers being dealt with in the game.

+++ Assets/Scripts/StateMachines/Customer/CustomerStateMachine.cs
@@ -43,7 +43,7 @@ public class CustomerStateMachine : StateMachine
 
     public void Pay()
     {
-        wallet.AddGold(order.orderItem.GetCost());
+        wallet.AddGold((double)order.orderItem.GetCost());
     }
 
     public void CustomerExit()
